code refactor, include query song setpoint fxn, new starter code
